Fixed the map to draw a circle with coordinates and radius ready to be used for location calculation
Fixed the map to draw a single circle that has a radius that is able to
be resized and dragged around to different places. Click the Save Area
button to add to database (still needs to add data to database). Also
made the windows larger because windows wasn't liking the smaller size.